+++ lib/Classes/track.dart
@@ -1,0 +1,6 @@
+import 'package:spoplusplusfy/Classes/voice.dart';
+
+class Track extends Voice {
+  Track({required super.audio, required super.id, required super.duration, required super.volume});
+
+}
