+++ pubspec.yaml
@@ -38,6 +38,8 @@ dependencies:
   flutter_svg: ^2.0.10+1
   just_audio: ^0.9.38
   sqflite: ^2.3.3+1
+  http: ^1.2.1
+  path_provider: ^2.0.11
 
 dev_dependencies:
   flutter_test:
